Fix Delpoyment Error
Status Quo: The template registration system was making the assumption
that all expected diretories exists in the build environment. So it does
not acount for a scenario where a required project directory does not
exist.

Git treats empty directories as non-existend. So whenever a "required
project directory" does not have any content, then it effectively does
not exist.

Problem: Whenever we had a project directory such as
 <project root>/sr/page/index/component that was empty, it was as if the
directory did not exist so the template registration system chokes on
it.

Solution: Update the template registration system to effectively handle
non-existent directories.