Fix Navbar Toggle Issue
Status Quo: When the navbar is toggled and the screen size changes to
"tablet width", the scrolling on the page remains disabled.

Problem: We need to be able to scroll when the screen size changes to
"tablet width" regardless of whether the navbar is toggled when the
change happens.

Solution: Add an event handler that disables the navbar whenever the
screen size changes to "tablet width".