Move Navbar Toggle Functionality To Navbar/Action
Status Quo: The functionality for toggling the Navbar is currently found
in the Navbar.js file.

Problem: By convention we want to have all functions operating on a
model to be located in the Action submodule, That way they can be easily
accessed by the models "update function"

Solution: move to toggle functions to the Action subdirectory.