Create "Navbar.patch" Function
Status Quo: Updating a Navbar object is done by directly assigning to
the Navbars attributes.

Problem: our convention dictates that we use a patch function to perform
updates to a model. This layer of abstraction allows us to change the
underlying data structure used to represent the model without having to
update consumers of the Navbar model. Not having this abstraction on our
Navbar object means that our object does not have the flexibility.

Solution: create a Navbar.patch function to perform updates.