Make Navbar Store Reference To It's Click Handlers
Status Quo: We want to be able to call DataStore.Action.execute on
the top-level data model from within the click handlers for a navbar.
There is currently no way for the click handlers to know what the
top-level data model will be at initialization time, much less what
the correct action to call should be.

Problem: We need a way to pass the correct top-level DataStore.Action to
the navbar click handlers at initialization time.

Solution: Refactor the Navbar model to store a reference to the click
handlers needed. Also refactor all Navbar.Action.Type.Initialize calls
to accept the click handlers to be passed to the model.