Refactor Navbar To Confirm To DataStore Model
Status Quo: I use a different system to toggle the navbar from the one
used to toggle the modal.

Problem: While both systems work, it's better to use a single system to
make things easier to understand.

Solution: Refactor Navbar to use the same system to manage state and
behaviour as that used by the Modal.