Refactor Navbar Model To Store DOM Element Id
Status Quo: We are using a constant to store the selector for the
element to be used as the navbar.

Problem: This is not ideal as it removes the flexibility of setting the
id at run time.

Solution: Refactor the Navbar model to store the Id for the element that
will be used to render the navbar.